+++ packages/auth/src/providers/cognito/utils/signUpHelpers.ts
@@ -10,8 +10,7 @@ import { AutoSignInCallback } from '../../../types/models';
 import { AuthError } from '../../../errors/AuthError';
 import { resetAutoSignIn, setAutoSignIn } from '../apis/autoSignIn';
 import { AUTO_SIGN_IN_EXCEPTION } from '../../../errors/constants';
-
-import { SignUpCommandOutput } from './clients/CognitoIdentityProvider/types';
+import { SignUpCommandOutput } from '../../../foundation/factories/serviceClients/cognitoIdentityProvider/types';
 
 const MAX_AUTOSIGNIN_POLLING_MS = 3 * 60 * 1000;