+++ packages/storage/src/providers/s3/apis/internal/list.ts
@@ -21,8 +21,8 @@ import {
 	validateStorageOperationInputWithPrefix,
 } from '../../utils';
 import {
-	ListAllOptionsWithPath,
-	ListPaginateOptionsWithPath,
+	ListAllWithPathOptions,
+	ListPaginateWithPathOptions,
 	ResolvedS3Config,
 } from '../../types/options';
 import {
@@ -274,7 +274,7 @@ const mapCommonPrefixesToExcludedSubpaths = (
 };
 
 const getDelimiter = (
-	options?: ListAllOptionsWithPath | ListPaginateOptionsWithPath,
+	options?: ListAllWithPathOptions | ListPaginateWithPathOptions,
 ): string | undefined => {
 	if (options?.subpathStrategy?.strategy === 'exclude') {
 		return options?.subpathStrategy?.delimiter ?? DEFAULT_DELIMITER;